refactor: centralize pack configuration
