+++ client/vite.config.js
@@ -3,6 +3,11 @@ import react from '@vitejs/plugin-react'
 import { VitePWA } from 'vite-plugin-pwa'
 
 export default defineConfig({
+  server: {
+    fs: {
+      allow: ['..']
+    }
+  },
   plugins: [
     react(),
     VitePWA({
