+++ shared/packs.js
@@ -1,8 +1,7 @@
-// On importe les listes de données statiques
-import europeanMushrooms from './common_european_mushrooms.json';
-import europeanTrees from './common_european_trees.json';
+const europeanMushrooms = require('../client/src/common_european_mushrooms.json');
+const europeanTrees = require('../client/src/common_european_trees.json');
 
-export const PACKS = [
+const PACKS = [
   {
     id: 'custom',
     type: 'custom', // Un type spécial pour notre filtre avancé
@@ -12,14 +11,14 @@ export const PACKS = [
   {
     id: 'european_mushrooms',
     type: 'list', // Ce pack est basé sur une liste d'IDs
-    title: 'Champignons commestibles d\'europe',
+    title: "Champignons commestibles d'europe",
     description: 'Une sélection des champignons les plus communs en Europe.',
     taxa_ids: europeanMushrooms.map(m => m.inaturalist_id)
   },
   {
     id: 'european_trees',
     type: 'list',
-    title: 'Arbres communs d\'europe',
+    title: "Arbres communs d'europe",
     description: 'Une sélection des arbres les plus communs en Europe.',
     taxa_ids: europeanTrees.map(t => t.inaturalist_id)
   },
@@ -27,7 +26,7 @@ export const PACKS = [
     id: 'world_birds',
     type: 'dynamic', // Ce pack sera interprété par le backend
     title: 'Oiseaux du monde',
-    description: 'Les 100 espèces d\'oiseaux les plus observées sur iNaturalist.',
+    description: "Les 100 espèces d'oiseaux les plus observées sur iNaturalist.",
     api_params: { taxon_id: '3', popular: 'true' }
   },
   {
@@ -56,17 +55,19 @@ export const PACKS = [
     type: 'dynamic',
     title: 'Flore méditerranéenne',
     description: 'Les plantes, arbres et fleurs typiques du bassin méditerranéen.',
-    api_params: { taxon_id: '47126', place_id: '53832', popular: 'true',}
+    api_params: { taxon_id: '47126', place_id: '53832', popular: 'true' }
   },
   {
-  id: 'great_barrier_reef_life',
-  type: 'dynamic',
-  title: 'Vie marine de la grande barrière de corail',
-  description: 'Poissons, coraux et mollusques du plus grand récif corallien du monde.',
-  api_params: {
-    taxon_id: '1',      // ID pour le règne Animalia (Animaux)
-    place_id: '131021',  // ID iNaturalist pour le Parc marin de la Grande Barrière de Corail
-    popular: 'true',
-  }
-},
+    id: 'great_barrier_reef_life',
+    type: 'dynamic',
+    title: 'Vie marine de la grande barrière de corail',
+    description: 'Poissons, coraux et mollusques du plus grand récif corallien du monde.',
+    api_params: {
+      taxon_id: '1',      // ID pour le règne Animalia (Animaux)
+      place_id: '131021',  // ID iNaturalist pour le Parc marin de la Grande Barrière de Corail
+      popular: 'true',
+    }
+  },
 ];
+
+module.exports = PACKS;
